FoundationDB.Error.Internal

Documentation

fdbEither :: MonadIO m => m (CFDBError, a) -> m (Either Error a)

fdbExcept :: (MonadError Error m, MonadIO m) => IO (CFDBError, a) -> m a

fdbEither' :: MonadIO m => m CFDBError -> m (Either Error ())

fdbExcept' :: (MonadError Error m, MonadIO m) => IO CFDBError -> m ()

liftFDBError :: MonadError Error m => Either CFDBError a -> m a

fdbThrowing :: IO (CFDBError, a) -> IO a

fdbThrowing' :: IO CFDBError -> IO ()

data Error

Represents all errors that can occur when running a Transaction.

Constructors

CError CError

Error FDBHsError

data DirLayerUserError

Constructors

CannotOpenRoot
Thrown if the user attempts to open the root directory.

PrefixInUse
Thrown if the user specifies a manual prefix that is already in use.

ManualPrefixConflict ByteString
Thrown if a prefix manually specified by the user previously conflicts with a prefix chosen by the automatic allocator. Includes the conflicting prefix.

LayerMismatch ByteString ByteString
The layer bytestring provided to open' does not match the layer already present. The mismatched layers are included in this constructor.

VersionError Word32 Word32 Word32
Thrown if the directory layer structure already in FoundationDB is a newer major version than that provided by this library. This would mean that the directory layer was originally created by a newer version of one of the FoundationDB client libraries. The major, minor, micro version of the directory layer are provided to this constructor.

data FDBHsError

Errors arising from the foundationdb-haskell library implementation.

Constructors

DirLayerUserError DirLayerUserError
Errors that can occur from user error when using the directory layer.

DirectoryLayerInternalError String
Errors that can occur when doing directory layer operations. These can be indicative of bugs in foundationdb-haskell.

ParseError String
Errors in parsing tuples.

MaxRetriesExceeded Error
Thrown by foundationdb-haskell's transaction retry logic. Contains the underlying error from the C bindings that caused the transaction to be retried.

UnsupportedAPIVersion
Thrown by foundationdb-haskell on startup if the Haskell code doesn't support the desired API version. This can happen even if the underlying C library does support the desired version -- we sometimes drop support for older versions sooner than the C API.

data CError

Errors that can come from the underlying C library. Most error names are self-explanatory. See https://apple.github.io/foundationdb/api-error-codes.html#developer-guide-error-codes for a description of these errors.

Constructors

OperationFailed

TimedOut

TransactionTooOld

FutureVersion

NotCommitted
Returned if a transaction failed because of a conflict.

CommitUnknownResult

TransactionCanceled

TransactionTimedOut

TooManyWatches

WatchesDisabled

AccessedUnreadable

DatabaseLocked

ClusterVersionChanged

ExternalClientAlreadyLoaded

OperationCancelled

FutureReleased

PlatformError

LargeAllocFailed

PerformanceCounterError

IOError

FileNotFound

BindFailed

FileNotReadable

FileNotWritable

NoClusterFileFound

FileTooLarge

ClientInvalidOperation

CommitReadIncomplete

TestSpecificationInvalid

KeyOutsideLegalRange

InvertedRange

InvalidOptionValue

InvalidOption

NetworkNotSetup

NetworkAlreadySetup

ReadVersionAlreadySet

VersionInvalid

RangeLimitsInvalid

InvalidDatabaseName

AttributeNotFound

FutureNotSet

FutureNotError

UsedDuringCommit

InvalidMutationType

TransactionInvalidVersion

NoCommitVersion

EnvironmentVariableNetworkOptionFailed

TransactionReadOnly

IncompatibleProtocolVersion

TransactionTooLarge

KeyTooLarge

ValueTooLarge

ConnectionStringInvalid

AddressInUse

InvalidLocalAddress

TLSError

UnsupportedOperation

APIVersionUnset

APIVersionAlreadySet

APIVersionInvalid

APIVersionNotSupported

ExactModeWithoutLimits

UnknownError

InternalError

OtherError

Fields

getOtherError :: CFDBError

toError :: CFDBError -> Maybe CError

Convert error int to CError sum. If 0 (which indicates success), returns Nothing. See isError for another way to guard for success.

toCFDBError :: CError -> CFDBError

retryable :: Error -> Bool

Returns true if the given error indicates that the erroring transaction can be tried, but the transaction might have already been committed. The transaction should only be retried if it is idempotent.

retryableNotCommitted :: Error -> Bool

Returns true if the given error indicates that the transaction was definitely not committed.