FoundationDB.Internal.Bindings

Description

Raw bindings to the underlying C client API. These are not memory safe. For documentation, see https://apple.github.io/foundationdb/api-c.html.

API versioning

currentAPIVersion :: Int

Current version of the installed FDB library. For example, returns 630 if you are using FoundationDB client 6.3.x.

selectAPIVersion :: Int -> IO CFDBError

Network

setupNetwork :: IO CFDBError

runNetwork :: IO CFDBError

stopNetwork :: IO CFDBError

networkSetOption :: NetworkOption -> IO CFDBError

Future

newtype Future a

Constructors

Future (Ptr (Future a))

futureCancel :: Future a -> IO ()

futureDestroy :: Future a -> IO ()

futureDestroyPtr :: FunPtr (Ptr a -> IO ())

futureBlockUntilReady :: Future a -> IO CFDBError

futureIsReady :: Future a -> IO Bool

futureReleaseMemory :: Future a -> IO ()

futureGetError :: Future a -> IO CFDBError

futureGetInt64 :: Future Int64 -> IO (CFDBError, Int64)

futureGetKey :: Future ByteString -> IO (Either CFDBError ByteString)

futureGetValue :: Future (Maybe ByteString) -> IO (Either CFDBError (Maybe ByteString))

futureGetStringArray :: Future [ByteString] -> IO (Either CFDBError [ByteString])

data FDBKeyValue

Constructors

FDBKeyValue

Fields

key :: Ptr ()

key_length :: Int

value :: Ptr ()

value_length :: Int

futureGetKeyValueArray :: Future [(ByteString, ByteString)] -> IO (Either CFDBError ([(ByteString, ByteString)], Bool))

Database

data DatabasePtr

Handle to the underlying C API client state.

createDatabase :: String -> IO (CFDBError, DatabasePtr)

databaseDestroy :: DatabasePtr -> IO ()

databaseSetOption :: DatabasePtr -> DatabaseOption -> IO CFDBError

databaseCreateTransaction :: DatabasePtr -> IO (CFDBError, Transaction)

Transaction

newtype Transaction

Constructors

Transaction (Ptr Transaction)

data KeySelector

Specifies a key in the database. See the official docs for more information. These can be supplied to getKey or used to build a Range.

Constructors

LastLessThan ByteString
Selects the lexicographically greatest key less than the specified key.

LastLessOrEq ByteString
Selects the lexicographically greatest less than or equal to the specified key.

FirstGreaterThan ByteString
Selects the lexicographically least key greater than the specified key.

FirstGreaterOrEq ByteString
Selects the lexicographically least key greater than or equal to the specified key.

WithOffset Int KeySelector
offsets a key selector. Using offset is preferred, since it handles normalization to prevent nested WithOffsets.

keySelectorBytes :: KeySelector -> ByteString

keySelectorTuple :: KeySelector -> (ByteString, Bool, Int)

Convert a KeySelector to its or_equal, offset settings. Equivalent to the macros FDB_KEYSEL_LAST_LESS_THAN etc.

tupleKeySelector :: (ByteString, Bool, Int) -> KeySelector

Inverse of keySelectorTuple

transactionDestroy :: Transaction -> IO ()

transactionDestroyPtr :: FunPtr (Ptr a -> IO ())

transactionSetOption :: Transaction -> TransactionOption -> IO CFDBError

transactionSetReadVersion :: Transaction -> Int64 -> IO ()

transactionGetReadVersion :: Transaction -> IO (Future Int64)

transactionGet :: Transaction -> ByteString -> Bool -> IO (Future (Maybe ByteString))

transactionGetKey :: Transaction -> ByteString -> Bool -> Int -> Bool -> IO (Future ByteString)

transactionGetAddressesForKey :: Transaction -> ByteString -> IO (Future [ByteString])

transactionGetRange

Arguments

:: Transaction

-> KeySelector
begin

-> KeySelector
end

-> Int
max number of pairs to return

-> Int
max number of bytes to return

-> FDBStreamingMode

-> Int
if FDBStreamingMode is FdbStreamingModeIterator, start this at 1 and increment by for each successive call reading this range. Otherwise, ignored.

-> Bool
isSnapshotRead

-> Bool
whether to return pairs in reverse order

-> IO (Future [(ByteString, ByteString)])

data FDBStreamingMode

Constructors

StreamingModeWantAll

StreamingModeIterator

StreamingModeExact

StreamingModeSmall

StreamingModeMedium

StreamingModeLarge

StreamingModeSerial

transactionSet :: Transaction -> ByteString -> ByteString -> IO ()

transactionClear :: Transaction -> ByteString -> IO ()

transactionClearRange :: Transaction -> ByteString -> ByteString -> IO ()

transactionAtomicOp

Arguments

:: Transaction

-> ByteString
key

-> MutationType

-> IO ()

transactionCommit :: Transaction -> IO (Future ())

transactionGetCommittedVersion :: Transaction -> IO (CFDBError, Int)

transactionGetApproximateSize :: Transaction -> IO (Future Int64)

transactionGetVersionstamp :: Transaction -> IO (Future ByteString)

transactionWatch :: Transaction -> ByteString -> IO (Future ())

transactionOnError :: Transaction -> CFDBError -> IO (Future ())

transactionReset :: Transaction -> IO ()

transactionCancel :: Transaction -> IO ()

transactionAddConflictRange :: Transaction -> ByteString -> ByteString -> FDBConflictRangeType -> IO CFDBError

data FDBConflictRangeType

Constructors

ConflictRangeTypeRead

ConflictRangeTypeWrite

Error

newtype CFDBError

Constructors

CFDBError

Fields

getCFDBError :: CInt

isError :: CFDBError -> Bool

Return True iff CFDBError value is an error (non-zero).

getError :: CFDBError -> IO String

errorPredicate :: FDBErrorPredicate -> CFDBError -> Bool

data FDBErrorPredicate

Constructors

ErrorPredicateRetryable

ErrorPredicateMaybeCommitted

ErrorPredicateRetryableNotCommitted

data MappedKeyValue

The result of a mapped key value query, containing the parent key and value (i.e., the key and value in the secondary index), and the range of key-value pairs retrieved based on the range query generated from the mapper's operation on the parent key and value.

WARNING: FDB's docs warn that the returned KeySelector fields have not yet been tested. Use them at your own risk!

Constructors

MappedKeyValue

Fields

parentKey :: ByteString

parentValue :: ByteString

mappedKeySelectorBegin :: KeySelector

mappedKeySelectorEnd :: KeySelector

mappedKeyValues :: [(ByteString, ByteString)]

getMappedRangeResult :: Future [MappedKeyValue] -> IO (Either CFDBError ([MappedKeyValue], Bool))

Extracts the mappedKeyValue results from the future, along with a bool indicating whether more results are available.

transactionGetMappedRange

Arguments

:: Transaction

-> KeySelector
begin

-> KeySelector
end

-> ByteString
Mapper name

-> Int
max number of pairs to return

-> Int
max number of bytes to return

-> FDBStreamingMode

-> Int
if FDBStreamingMode is FdbStreamingModeIterator, start this at 1 and increment by for each successive call reading this range. Otherwise, ignored.

-> Bool
isSnapshotRead

-> Bool
whether to return pairs in reverse order

-> IO (Future [MappedKeyValue])