FoundationDB.Internal.Database

Documentation

data FoundationDBOptions

Options set at the connection level for FoundationDB.

Constructors

FoundationDBOptions

Fields

apiVersion :: Int

Desired API version. See currentAPIVersion for the latest version installed on your system. The C API (and this library) allow you to choose any version earlier than currentAPIVersion to get the client behavior of that version of the FoundationDB client library.

clusterFile :: Maybe FilePath

Path to your fdb.cluster file. If Nothing, uses default location.

networkOptions :: [NetworkOption]

Additional network options. Each will be set in order.

databaseOptions :: [DatabaseOption]

Additional database options. Each will be set in order.

defaultOptions :: FoundationDBOptions

Uses the current API version, the default cluster file location, and no additional options.

data Database

Constructors

Database

Fields

databasePtr :: DatabasePtr

databaseFoundationDBOptions :: FoundationDBOptions

apiVersionInUse :: Database -> Int

Returns the API version that was specified in the apiVersion field when the FDB client was initialized.