FoundationDB.Layer.Directory

Description

The directory layer. The directory layer provides tools for creating a hierarchy of Subspaces, which can be operated on efficiently with a directory-like API.

This implementation of the directory layer does not yet support directory partitions. They will be added in the future.

Documentation

data DirectoryLayer

Represents a directory tree. A value of this type must be supplied to all functions in this module.

defaultDirLayer :: DirectoryLayer

The default directory layer has node subspace prefix 0xfe. This corresponds to using the defaults for all arguments to the DirectoryLayer constructor in other languages' bindings.

newDirectoryLayer

Arguments

:: Subspace
node subspace for directory metadata

-> Subspace
content subspace

-> Bool
allow manual prefixes

-> DirectoryLayer

Creates a new directory layer, containing a hierarchy of directories.

data Directory

Represents a single directory.

type Path = [Text]

A path is a list of unicode strings.

dirSubspace :: Directory -> Subspace

Gets the content subspace of a directory, which can be used to store tuple-based keys.

dirPath :: Directory -> Path

Gets the path of a directory.

dirLayer :: Directory -> ByteString

Gets the layer tag that was specified when the directory was created.

open :: DirectoryLayer -> Path -> Transaction (Maybe Directory)

Tries to open a directory at the given path. If the directory exists, returns it. Otherwise, returns Nothing.

createOrOpen :: DirectoryLayer -> Path -> Transaction Directory

Opens a directory at the given path. If the directory does not exist, it is created.

move

Arguments

:: DirectoryLayer

-> Path
from path

-> Path
to path

-> Transaction (Maybe MoveError)

Move a directory from one path to another.

remove :: DirectoryLayer -> Path -> Transaction Bool

Remove a directory path, its contents, and all subdirectories. Returns True if removal succeeds. Fails for nonexistent paths and the root directory.

exists :: DirectoryLayer -> Path -> Transaction Bool

Returns True iff the given path exists.

list :: DirectoryLayer -> Path -> Transaction (Seq Text)

List the names of the immediate subdirectories of a directory. Returns an empty list if the directory does not exist.

Advanced usage

open'

Arguments

:: DirectoryLayer

-> Path

-> ByteString
layer

-> Maybe ByteString
optional custom prefix

-> Transaction (Maybe Directory)

Open a directory, with optional custom prefix and layer. Returns Nothing if the directory doesn't exist.

createOrOpen'

Arguments

:: DirectoryLayer

-> Path

-> ByteString
layer

-> Maybe ByteString
optional custom prefix

-> Transaction Directory

Opens a directory at the given path, with optional custom prefix and layer.