FoundationDB.Layer.Tuple.Internal

Documentation

data Elem

Elements of tuples. A tuple is represented as a list of these. Note that a tuple may contain at most one incomplete version stamp. Future versions of this library may introduce a more strongly typed tuple representation that enforces this restriction.

Constructors

None
Corresponds to null or nil types in other language bindings.

Tuple [Elem]
Nested tuples.

Bytes ByteString

Text Text

Int Integer
Variable-length integer encodings. For values that fit within a 64-bit signed integer, the standard integer encoding is used. For larger values, the provisional spec for Java and Python values is used.

Float Float

Double Double

Bool Bool

UUID Word32 Word32 Word32 Word32
Crude UUID to avoid dependency on UUID library. Interconvertible with toWords and fromWords in UUID.

CompleteVS (Versionstamp 'Complete)

IncompleteVS (Versionstamp 'Incomplete)
This constructor is to be used in conjunction with encodeTupleElems and the setVersionstampedKey atomic operation. See encodeTupleElems for more information.

sizeLimits :: Array Int Integer

bisectSize :: Integer -> Int

Returns smallest size limit greater than input.

bitLen :: Integer -> Int

Returns the minimum number of bits needed to encode the given int.

byteLen :: Integer -> Int

Returns the minimum number of bytes needed to encode the given int.

nullCode :: Word8

bytesCode :: Word8

stringCode :: Word8

nestedCode :: Word8

zeroCode :: Word8

posEndCode :: Word8

negStartCode :: Word8

floatCode :: Word8

doubleCode :: Word8

falseCode :: Word8

trueCode :: Word8

uuidCode :: Word8

versionstampCode :: Word8

data SerializationState

Constructors

SerializationState

Fields

currLength :: Int

incompleteVersionstampPos :: Maybe Int

newtype PutTuple a

Constructors

PutTuple

Fields

unPutTuple :: StateT SerializationState PutM a

runPutTuple :: PutTuple () -> (ByteString, Maybe Int)

returns the serialized tuple and the position of the incomplete version stamp, if any.

incrLength :: Int -> PutTuple ()

liftPutM :: PutM a -> PutTuple a

putWord8 :: Word8 -> PutTuple ()

putWord16be :: Word16 -> PutTuple ()

putWord32be :: Word32 -> PutTuple ()

putWord64be :: Word64 -> PutTuple ()

putByteString :: ByteString -> PutTuple ()

encodeBytes :: ByteString -> PutTuple ()

truncatedInt :: Int -> Integer -> ByteString

encodeLargePosInt :: Integer -> PutTuple ()

encodeLargeNegInt :: Integer -> PutTuple ()

encodePosInt :: Integer -> PutTuple ()

encodeNegInt :: Integer -> PutTuple ()

floatAdjust :: Bool -> ByteString -> ByteString

given an IEEE 754 float/double, adjust it for encoding.

encodeElem

Arguments

:: Bool
Whether we are inside a nested tuple

-> Elem
elem to encode

-> PutTuple ()

encodeTupleElems :: Traversable t => t Elem -> ByteString

Encodes a tuple from a list of tuple elements. Returns the encoded tuple.

Warning: this function can throw an Error with TupleIntTooLarge if you pass an Int element that requires more than 255 bytes to serialize. Since the smallest such number is 614 decimal digits long, we deemed this situation unlikely enough that it wasn't worth returning a sum type from this function.

Note: this encodes to the format expected by FoundationDB as input, which is slightly different from the format returned by FoundationDB as output. The difference is that if the encoded bytes include an incomplete version stamp, four bytes are appended to the end to indicate the index of the incomplete version stamp so that FoundationDB can fill in the transaction version and batch order when this function is used in conjunction with setVersionstampedKey and setVersionstampedValue:

do let k = pack mySubspace [IncompleteVS (IncompleteVersionstamp 123)]
   atomicOp k (setVersionstampedKey "my_value")

Because FoundationDB uses two bytes at the end of the key for this, only one IncompleteVS can be used per key.

This also means that (decodeTupleElems . encodeTupleElems) gives strange results when an IncompleteVS is present in the input, because the two extra bytes are interpreted as being part of the tuple.

>>> decodeTupleElems $ encodeTupleElems [IncompleteVS (IncompleteVersionstamp 1)]
Right [IncompleteVS (IncompleteVersionstamp 1),Bytes "",None,None]

For this reason, decodeTupleElems should only be called on keys that have been returned from the database, because setVersionstampedKey drops the last two bytes when it writes the key to the database.

encodeTupleElemsWPrefix :: Traversable t => ByteString -> t Elem -> ByteString

Like encodeTupleElems, but prepends a raw bytestring prefix to the tuple. This is used by the subspace and directory layers.

decodeTupleElems :: ByteString -> Either String [Elem]

Decodes a tuple, or returns a parse error. This function will never return IncompleteVS tuple elements. See the note on encodeTupleElems for more information.

decodeTupleElemsWPrefix

Arguments

:: ByteString
expected prefix

-> ByteString
encoded tuple

-> Either String [Elem]

Decodes a tuple that was encoded with a given prefix. Fails if the input prefix is not actually a prefix of the encoded tuple.

runGetComplete :: ByteString -> Get a -> Either String a

decodeElem :: Bool -> Get Elem

expectCode :: Word8 -> Get ()

bytesTerminator :: Get ()

getBytesUntilTerminator :: Get ByteString

Reads all bytes up to (but not including) the terminator byte

decodeBytesElem :: Get Elem

decodeTextElem :: Get Elem

decodeSmallPosInt :: Word8 -> Get Elem

decodeSmallNegInt :: Word8 -> Get Elem

decodeLargeNegInt :: Get Elem

decodeLargePosInt :: Get Elem

decodeFloatElem :: Get Elem

decodeDoubleElem :: Get Elem

decodeUUIDElem :: Get Elem

decodeTupleElem :: Get Elem

decodeVersionstamp :: Get Elem