FoundationDB.Options

Description

NOTE: This file is generated from fdb.options by the generate-options executable in this project. All documentation on the individual options in this namespace comes from FoundationDB's documentation in fdb.options.

Documentation

data NetworkOption

Constructors

NetworkOptionString Int String

NetworkOptionInt Int Int

NetworkOptionBytes Int ByteString

NetworkOptionFlag Int

localAddress :: String -> NetworkOption

Deprecated: Deprecated in FDB C API

Deprecated

clusterFile :: String -> NetworkOption

Deprecated: Deprecated in FDB C API

Deprecated

traceEnable :: String -> NetworkOption

Enables trace output to a file in a directory of the clients choosing

traceRollSize :: Int -> NetworkOption

Sets the maximum size in bytes of a single trace output file. This value should be in the range ``[0, INT64_MAX]``. If the value is set to 0, there is no limit on individual file size. The default is a maximum size of 10,485,760 bytes.

traceMaxLogsSize :: Int -> NetworkOption

Sets the maximum size of all the trace output files put together. This value should be in the range ``[0, INT64_MAX]``. If the value is set to 0, there is no limit on the total size of the files. The default is a maximum size of 104,857,600 bytes. If the default roll size is used, this means that a maximum of 10 trace files will be written at a time.

traceLogGroup :: String -> NetworkOption

Sets the LogGroup attribute with the specified value for all events in the trace output files. The default log group is 'default'.

traceFormat :: String -> NetworkOption

Select the format of the log files. xml (the default) and json are supported.

knob :: String -> NetworkOption

Set internal tuning or debugging knobs

tlsPlugin :: String -> NetworkOption

Deprecated: Deprecated in FDB C API

Deprecated

tlsCertBytes :: ByteString -> NetworkOption

Set the certificate chain

tlsCertPath :: String -> NetworkOption

Set the file from which to load the certificate chain

tlsKeyBytes :: ByteString -> NetworkOption

Set the private key corresponding to your own certificate

tlsKeyPath :: String -> NetworkOption

Set the file from which to load the private key corresponding to your own certificate

tlsVerifyPeers :: ByteString -> NetworkOption

Set the peer certificate field verification criteria

buggifyEnable :: NetworkOption

buggifyDisable :: NetworkOption

buggifySectionActivatedProbability :: Int -> NetworkOption

Set the probability of a BUGGIFY section being active for the current execution. Only applies to code paths first traversed AFTER this option is changed.

buggifySectionFiredProbability :: Int -> NetworkOption

Set the probability of an active BUGGIFY section being fired

tlsCaBytes :: ByteString -> NetworkOption

Set the ca bundle

tlsCaPath :: String -> NetworkOption

Set the file from which to load the certificate authority bundle

tlsPassword :: String -> NetworkOption

Set the passphrase for encrypted private key. Password should be set before setting the key for the password to be used.

disableMultiVersionClientApi :: NetworkOption

Disables the multi-version client API and instead uses the local client directly. Must be set before setting up the network.

callbacksOnExternalThreads :: NetworkOption

If set, callbacks from external client libraries can be called from threads created by the FoundationDB client library. Otherwise, callbacks will be called from either the thread used to add the callback or the network thread. Setting this option can improve performance when connected using an external client, but may not be safe to use in all environments. Must be set before setting up the network. WARNING: This feature is considered experimental at this time.

externalClientLibrary :: String -> NetworkOption

Adds an external client library for use by the multi-version client API. Must be set before setting up the network.

externalClientDirectory :: String -> NetworkOption

Searches the specified path for dynamic libraries and adds them to the list of client libraries for use by the multi-version client API. Must be set before setting up the network.

disableLocalClient :: NetworkOption

Prevents connections through the local client, allowing only connections through externally loaded client libraries. Intended primarily for testing.

disableClientStatisticsLogging :: NetworkOption

Disables logging of client statistics, such as sampled transaction activity.

enableSlowTaskProfiling :: NetworkOption

Enables debugging feature to perform slow task profiling. Requires trace logging to be enabled. WARNING: this feature is not recommended for use in production.

supportedClientVersions :: String -> NetworkOption

This option is set automatically to communicate the list of supported clients to the active client.

externalClient :: NetworkOption

This option is set automatically on all clients loaded externally using the multi-version API.

externalClientTransportId :: Int -> NetworkOption

This option tells a child on a multiversion client what transport ID to use.

data DatabaseOption

Constructors

DatabaseOptionString Int String

DatabaseOptionInt Int Int

DatabaseOptionBytes Int ByteString

DatabaseOptionFlag Int

locationCacheSize :: Int -> DatabaseOption

Set the size of the client location cache. Raising this value can boost performance in very large databases where clients access data in a near-random pattern. Defaults to 100000.

maxWatches :: Int -> DatabaseOption

Set the maximum number of watches allowed to be outstanding on a database connection. Increasing this number could result in increased resource usage. Reducing this number will not cancel any outstanding watches. Defaults to 10000 and cannot be larger than 1000000.

machineId :: String -> DatabaseOption

Specify the machine ID that was passed to fdbserver processes running on the same machine as this client, for better location-aware load balancing.

datacenterId :: String -> DatabaseOption

Specify the datacenter ID that was passed to fdbserver processes running in the same datacenter as this client, for better location-aware load balancing.

transactionTimeout :: Int -> DatabaseOption

Set a timeout in milliseconds which, when elapsed, will cause each transaction automatically to be cancelled. This sets the `timeout` option of each transaction created by this database. See the transaction option description for more information. Using this option requires that the API version is 610 or higher.

transactionRetryLimit :: Int -> DatabaseOption

Set a timeout in milliseconds which, when elapsed, will cause a transaction automatically to be cancelled. This sets the `retry_limit` option of each transaction created by this database. See the transaction option description for more information.

transactionMaxRetryDelay :: Int -> DatabaseOption

Set the maximum amount of backoff delay incurred in the call to `onError` if the error is retryable. This sets the `max_retry_delay` option of each transaction created by this database. See the transaction option description for more information.

dbSnapshotRywEnable :: DatabaseOption

Snapshot read operations will see the results of writes done in the same transaction. This is the default behavior.

dbSnapshotRywDisable :: DatabaseOption

Snapshot read operations will not see the results of writes done in the same transaction. This was the default behavior prior to API version 300.

data TransactionOption

Constructors

TransactionOptionString Int String

TransactionOptionInt Int Int

TransactionOptionBytes Int ByteString

TransactionOptionFlag Int

causalWriteRisky :: TransactionOption

The transaction, if not self-conflicting, may be committed a second time after commit succeeds, in the event of a fault

causalReadRisky :: TransactionOption

The read version will be committed, and usually will be the latest committed, but might not be the latest committed in the event of a fault or partition

causalReadDisable :: TransactionOption

nextWriteNoWriteConflictRange :: TransactionOption

The next write performed on this transaction will not generate a write conflict range. As a result, other transactions which read the key(s) being modified by the next write will not conflict with this transaction. Care needs to be taken when using this option on a transaction that is shared between multiple threads. When setting this option, write conflict ranges will be disabled on the next write operation, regardless of what thread it is on.

commitOnFirstProxy :: TransactionOption

Committing this transaction will bypass the normal load balancing across proxies and go directly to the specifically nominated 'first proxy'.

checkWritesEnable :: TransactionOption

readYourWritesDisable :: TransactionOption

Reads performed by a transaction will not see any prior mutations that occured in that transaction, instead seeing the value which was in the database at the transaction's read version. This option may provide a small performance benefit for the client, but also disables a number of client-side optimizations which are beneficial for transactions which tend to read and write the same keys within a single transaction.

readAheadDisable :: TransactionOption

Deprecated: Deprecated in FDB C API

Deprecated

durabilityDatacenter :: TransactionOption

durabilityRisky :: TransactionOption

durabilityDevNullIsWebScale :: TransactionOption

Deprecated: Deprecated in FDB C API

Deprecated

prioritySystemImmediate :: TransactionOption

Specifies that this transaction should be treated as highest priority and that lower priority transactions should block behind this one. Use is discouraged outside of low-level tools

priorityBatch :: TransactionOption

Specifies that this transaction should be treated as low priority and that default priority transactions will be processed first. Batch priority transactions will also be throttled at load levels smaller than for other types of transactions and may be fully cut off in the event of machine failures. Useful for doing batch work simultaneously with latency-sensitive work

initializeNewDatabase :: TransactionOption

This is a write-only transaction which sets the initial configuration. This option is designed for use by database system tools only.

accessSystemKeys :: TransactionOption

Allows this transaction to read and modify system keys (those that start with the byte 0xFF)

readSystemKeys :: TransactionOption

Allows this transaction to read system keys (those that start with the byte 0xFF)

debugDump :: TransactionOption

debugRetryLogging :: String -> TransactionOption

transactionLoggingEnable :: String -> TransactionOption

Deprecated: Deprecated in FDB C API

Deprecated

debugTransactionIdentifier :: String -> TransactionOption

Sets a client provided identifier for the transaction that will be used in scenarios like tracing or profiling. Client trace logging or transaction profiling must be separately enabled.

logTransaction :: TransactionOption

Enables tracing for this transaction and logs results to the client trace logs. The DEBUG_TRANSACTION_IDENTIFIER option must be set before using this option, and client trace logging must be enabled and to get log output.

timeout :: Int -> TransactionOption

Set a timeout in milliseconds which, when elapsed, will cause the transaction automatically to be cancelled. Valid parameter values are ``[0, INT_MAX]``. If set to 0, will disable all timeouts. All pending and any future uses of the transaction will throw an exception. The transaction can be used again after it is reset. Prior to API version 610, like all other transaction options, the timeout must be reset after a call to `onError`. If the API version is 610 or greater, the timeout is not reset after an `onError` call. This allows the user to specify a longer timeout on specific transactions than the default timeout specified through the `transaction_timeout` database option without the shorter database timeout cancelling transactions that encounter a retryable error. Note that at all API versions, it is safe and legal to set the timeout each time the transaction begins, so most code written assuming the older behavior can be upgraded to the newer behavior without requiring any modification, and the caller is not required to implement special logic in retry loops to only conditionally set this option.

retryLimit :: Int -> TransactionOption

Set a maximum number of retries after which additional calls to `onError` will throw the most recently seen error code. Valid parameter values are ``[-1, INT_MAX]``. If set to -1, will disable the retry limit. Prior to API version 610, like all other transaction options, the retry limit must be reset after a call to `onError`. If the API version is 610 or greater, the retry limit is not reset after an `onError` call. Note that at all API versions, it is safe and legal to set the retry limit each time the transaction begins, so most code written assuming the older behavior can be upgraded to the newer behavior without requiring any modification, and the caller is not required to implement special logic in retry loops to only conditionally set this option.

maxRetryDelay :: Int -> TransactionOption

Set the maximum amount of backoff delay incurred in the call to `onError` if the error is retryable. Defaults to 1000 ms. Valid parameter values are ``[0, INT_MAX]``. If the maximum retry delay is less than the current retry delay of the transaction, then the current retry delay will be clamped to the maximum retry delay. Prior to API version 610, like all other transaction options, the maximum retry delay must be reset after a call to `onError`. If the API version is 610 or greater, the retry limit is not reset after an `onError` call. Note that at all API versions, it is safe and legal to set the maximum retry delay each time the transaction begins, so most code written assuming the older behavior can be upgraded to the newer behavior without requiring any modification, and the caller is not required to implement special logic in retry loops to only conditionally set this option.

snapshotRywEnable :: TransactionOption

Snapshot read operations will see the results of writes done in the same transaction. This is the default behavior.

snapshotRywDisable :: TransactionOption

Snapshot read operations will not see the results of writes done in the same transaction. This was the default behavior prior to API version 300.

lockAware :: TransactionOption

The transaction can read and write to locked databases, and is resposible for checking that it took the lock.

usedDuringCommitProtectionDisable :: TransactionOption

By default, operations that are performed on a transaction while it is being committed will not only fail themselves, but they will attempt to fail other in-flight operations (such as the commit) as well. This behavior is intended to help developers discover situations where operations could be unintentionally executed after the transaction has been reset. Setting this option removes that protection, causing only the offending operation to fail.

readLockAware :: TransactionOption

The transaction can read from locked databases.

firstInBatch :: TransactionOption

No other transactions will be applied before this transaction within the same commit version.

useProvisionalProxies :: TransactionOption

This option should only be used by tools which change the database configuration.

data StreamingMode

Constructors

StreamingModeString Int String

StreamingModeInt Int Int

StreamingModeBytes Int ByteString

StreamingModeFlag Int

wantAll :: StreamingMode

Client intends to consume the entire range and would like it all transferred as early as possible.

iterator :: StreamingMode

The default. The client doesn't know how much of the range it is likely to used and wants different performance concerns to be balanced. Only a small portion of data is transferred to the client initially (in order to minimize costs if the client doesn't read the entire range), and as the caller iterates over more items in the range larger batches will be transferred in order to minimize latency.

exact :: StreamingMode

Infrequently used. The client has passed a specific row limit and wants that many rows delivered in a single batch. Because of iterator operation in client drivers make request batches transparent to the user, consider `WANT_ALL` StreamingMode instead. A row limit must be specified if this mode is used.

small :: StreamingMode

Infrequently used. Transfer data in batches small enough to not be much more expensive than reading individual rows, to minimize cost if iteration stops early.

medium :: StreamingMode

Infrequently used. Transfer data in batches sized in between small and large.

large :: StreamingMode

Infrequently used. Transfer data in batches large enough to be, in a high-concurrency environment, nearly as efficient as possible. If the client stops iteration early, some disk and network bandwidth may be wasted. The batch size may still be too small to allow a single client to get high throughput from the database, so if that is what you need consider the SERIAL StreamingMode.

serial :: StreamingMode

Transfer data in batches large enough that an individual client can get reasonable read bandwidth from the database. If the client stops iteration early, considerable disk and network bandwidth may be wasted.

data MutationType

Constructors

MutationTypeString Int String

MutationTypeInt Int Int

MutationTypeBytes Int ByteString

MutationTypeFlag Int

add :: ByteString -> MutationType

Performs an addition of little-endian integers. If the existing value in the database is not present or shorter than `param`, it is first extended to the length of `param` with zero bytes. If `param` is shorter than the existing value in the database, the existing value is truncated to match the length of `param`. The integers to be added must be stored in a little-endian representation. They can be signed in two's complement representation or unsigned. You can add to an integer at a known offset in the value by prepending the appropriate number of zero bytes to `param` and padding with zero bytes to match the length of the value. However, this offset technique requires that you know the addition will not cause the integer field within the value to overflow.

and :: ByteString -> MutationType

Deprecated: Deprecated in FDB C API

Deprecated

bitAnd :: ByteString -> MutationType

Performs a bitwise `and` operation. If the existing value in the database is not present, then `param` is stored in the database. If the existing value in the database is shorter than `param`, it is first extended to the length of `param` with zero bytes. If `param` is shorter than the existing value in the database, the existing value is truncated to match the length of `param`.

or :: ByteString -> MutationType

Deprecated: Deprecated in FDB C API

Deprecated

bitOr :: ByteString -> MutationType

Performs a bitwise `or` operation. If the existing value in the database is not present or shorter than `param`, it is first extended to the length of `param` with zero bytes. If `param` is shorter than the existing value in the database, the existing value is truncated to match the length of `param`.

xor :: ByteString -> MutationType

Deprecated: Deprecated in FDB C API

Deprecated

bitXor :: ByteString -> MutationType

Performs a bitwise `xor` operation. If the existing value in the database is not present or shorter than `param`, it is first extended to the length of `param` with zero bytes. If `param` is shorter than the existing value in the database, the existing value is truncated to match the length of `param`.

appendIfFits :: ByteString -> MutationType

Appends `param` to the end of the existing value already in the database at the given key (or creates the key and sets the value to `param` if the key is empty). This will only append the value if the final concatenated value size is less than or equal to the maximum value size (i.e., if it fits). WARNING: No error is surfaced back to the user if the final value is too large because the mutation will not be applied until after the transaction has been committed. Therefore, it is only safe to use this mutation type if one can guarantee that one will keep the total value size under the maximum size.

max :: ByteString -> MutationType

Performs a little-endian comparison of byte strings. If the existing value in the database is not present or shorter than `param`, it is first extended to the length of `param` with zero bytes. If `param` is shorter than the existing value in the database, the existing value is truncated to match the length of `param`. The larger of the two values is then stored in the database.

min :: ByteString -> MutationType

Performs a little-endian comparison of byte strings. If the existing value in the database is not present, then `param` is stored in the database. If the existing value in the database is shorter than `param`, it is first extended to the length of `param` with zero bytes. If `param` is shorter than the existing value in the database, the existing value is truncated to match the length of `param`. The smaller of the two values is then stored in the database.

setVersionstampedKey :: ByteString -> MutationType

Transforms `key` using a versionstamp for the transaction. Sets the transformed key in the database to `param`. The key is transformed by removing the final four bytes from the key and reading those as a little-Endian 32-bit integer to get a position `pos`. The 10 bytes of the key from `pos` to ``pos + 10`` are replaced with the versionstamp of the transaction used. The first byte of the key is position 0. A versionstamp is a 10 byte, unique, monotonically (but not sequentially) increasing value for each committed transaction. The first 8 bytes are the committed version of the database (serialized in big-Endian order). The last 2 bytes are monotonic in the serialization order for transactions. WARNING: At this time, versionstamps are compatible with the Tuple layer only in the Java, Python, and Go bindings. Also, note that prior to API version 520, the offset was computed from only the final two bytes rather than the final four bytes.

setVersionstampedValue :: ByteString -> MutationType

Transforms `param` using a versionstamp for the transaction. Sets the `key` given to the transformed `param`. The parameter is transformed by removing the final four bytes from `param` and reading those as a little-Endian 32-bit integer to get a position `pos`. The 10 bytes of the parameter from `pos` to ``pos + 10`` are replaced with the versionstamp of the transaction used. The first byte of the parameter is position 0. A versionstamp is a 10 byte, unique, monotonically (but not sequentially) increasing value for each committed transaction. The first 8 bytes are the committed version of the database (serialized in big-Endian order). The last 2 bytes are monotonic in the serialization order for transactions. WARNING: At this time, versionstamps are compatible with the Tuple layer only in the Java, Python, and Go bindings. Also, note that prior to API version 520, the versionstamp was always placed at the beginning of the parameter rather than computing an offset.

byteMin :: ByteString -> MutationType

Performs lexicographic comparison of byte strings. If the existing value in the database is not present, then `param` is stored. Otherwise the smaller of the two values is then stored in the database.

byteMax :: ByteString -> MutationType

Performs lexicographic comparison of byte strings. If the existing value in the database is not present, then `param` is stored. Otherwise the larger of the two values is then stored in the database.

compareAndClear :: ByteString -> MutationType

Performs an atomic ``compare and clear`` operation. If the existing value in the database is equal to the given value, then given key is cleared.

data ConflictRangeType

Constructors

ConflictRangeTypeString Int String

ConflictRangeTypeInt Int Int

ConflictRangeTypeBytes Int ByteString

ConflictRangeTypeFlag Int

read :: ConflictRangeType

Used to add a read conflict range

write :: ConflictRangeType

Used to add a write conflict range

data ErrorPredicate

Constructors

ErrorPredicateString Int String

ErrorPredicateInt Int Int

ErrorPredicateBytes Int ByteString

ErrorPredicateFlag Int

retryable :: ErrorPredicate

Returns `true` if the error indicates the operations in the transactions should be retried because of transient error.

maybeCommitted :: ErrorPredicate

Returns `true` if the error indicates the transaction may have succeeded, though not in a way the system can verify.

retryableNotCommitted :: ErrorPredicate

Returns `true` if the error indicates the transaction has not committed, though in a way that can be retried.