FoundationDB.Options.DatabaseOption

Description

NOTE: This file is generated from fdb.options by the generate-options executable in this project. All documentation on the individual options in this namespace comes from FoundationDB's documentation in fdb.options.

Documentation

data DatabaseOption

Constructors

DatabaseOptionString Int String

DatabaseOptionInt Int Int

DatabaseOptionBytes Int ByteString

DatabaseOptionFlag Int

locationCacheSize :: Int -> DatabaseOption

Set the size of the client location cache. Raising this value can boost performance in very large databases where clients access data in a near-random pattern. Defaults to 100000.

maxWatches :: Int -> DatabaseOption

Set the maximum number of watches allowed to be outstanding on a database connection. Increasing this number could result in increased resource usage. Reducing this number will not cancel any outstanding watches. Defaults to 10000 and cannot be larger than 1000000.

machineId :: String -> DatabaseOption

Specify the machine ID that was passed to fdbserver processes running on the same machine as this client, for better location-aware load balancing.

datacenterId :: String -> DatabaseOption

Specify the datacenter ID that was passed to fdbserver processes running in the same datacenter as this client, for better location-aware load balancing.

snapshotRywEnable :: DatabaseOption

Snapshot read operations will see the results of writes done in the same transaction. This is the default behavior.

snapshotRywDisable :: DatabaseOption

Snapshot read operations will not see the results of writes done in the same transaction. This was the default behavior prior to API version 300.

transactionLoggingMaxFieldLength :: Int -> DatabaseOption

Sets the maximum escaped length of key and value fields to be logged to the trace file via the LOG_TRANSACTION option. This sets the `transaction_logging_max_field_length` option of each transaction created by this database. See the transaction option description for more information.

transactionTimeout :: Int -> DatabaseOption

Set a timeout in milliseconds which, when elapsed, will cause each transaction automatically to be cancelled. This sets the `timeout` option of each transaction created by this database. See the transaction option description for more information. Using this option requires that the API version is 610 or higher.

transactionRetryLimit :: Int -> DatabaseOption

Set a maximum number of retries after which additional calls to `onError` will throw the most recently seen error code. This sets the `retry_limit` option of each transaction created by this database. See the transaction option description for more information.

transactionMaxRetryDelay :: Int -> DatabaseOption

Set the maximum amount of backoff delay incurred in the call to `onError` if the error is retryable. This sets the `max_retry_delay` option of each transaction created by this database. See the transaction option description for more information.

transactionSizeLimit :: Int -> DatabaseOption

Set the maximum transaction size in bytes. This sets the `size_limit` option on each transaction created by this database. See the transaction option description for more information.

transactionCausalReadRisky :: DatabaseOption

The read version will be committed, and usually will be the latest committed, but might not be the latest committed in the event of a simultaneous fault and misbehaving clock.

transactionIncludePortInAddress :: DatabaseOption

Addresses returned by get_addresses_for_key include the port when enabled. As of api version 630, this option is enabled by default and setting this has no effect.