FoundationDB.Options.NetworkOption

Description

by the generate-options executable in this project. All documentation on the individual options in this namespace comes from FoundationDB's documentation in fdb.options.

Documentation

data NetworkOption

Constructors

NetworkOptionString Int String

NetworkOptionInt Int Int

NetworkOptionBytes Int ByteString

NetworkOptionFlag Int

localAddress :: String -> NetworkOption

Deprecated: Deprecated in FDB C API

Deprecated

clusterFile :: String -> NetworkOption

Deprecated: Deprecated in FDB C API

Deprecated

traceEnable :: String -> NetworkOption

Enables trace output to a file in a directory of the clients choosing

traceRollSize :: Int -> NetworkOption

Sets the maximum size in bytes of a single trace output file. This value should be in the range ``[0, INT64_MAX]``. If the value is set to 0, there is no limit on individual file size. The default is a maximum size of 10,485,760 bytes.

traceMaxLogsSize :: Int -> NetworkOption

Sets the maximum size of all the trace output files put together. This value should be in the range ``[0, INT64_MAX]``. If the value is set to 0, there is no limit on the total size of the files. The default is a maximum size of 104,857,600 bytes. If the default roll size is used, this means that a maximum of 10 trace files will be written at a time.

traceLogGroup :: String -> NetworkOption

Sets the LogGroup attribute with the specified value for all events in the trace output files. The default log group is 'default'.

traceFormat :: String -> NetworkOption

Select the format of the log files. xml (the default) and json are supported.

traceClockSource :: String -> NetworkOption

Select clock source for trace files. now (the default) or realtime are supported.

traceFileIdentifier :: String -> NetworkOption

Once provided, this string will be used to replace the port/PID in the log file names.

tracePartialFileSuffix :: String -> NetworkOption

Set file suffix for partially written log files.

knob :: String -> NetworkOption

Set internal tuning or debugging knobs

tlsPlugin :: String -> NetworkOption

Deprecated: Deprecated in FDB C API

Deprecated

tlsCertBytes :: ByteString -> NetworkOption

Set the certificate chain

tlsCertPath :: String -> NetworkOption

Set the file from which to load the certificate chain

tlsKeyBytes :: ByteString -> NetworkOption

Set the private key corresponding to your own certificate

tlsKeyPath :: String -> NetworkOption

Set the file from which to load the private key corresponding to your own certificate

tlsVerifyPeers :: ByteString -> NetworkOption

Set the peer certificate field verification criteria

buggifyEnable :: NetworkOption

buggifyDisable :: NetworkOption

buggifySectionActivatedProbability :: Int -> NetworkOption

Set the probability of a BUGGIFY section being active for the current execution. Only applies to code paths first traversed AFTER this option is changed.

buggifySectionFiredProbability :: Int -> NetworkOption

Set the probability of an active BUGGIFY section being fired

tlsCaBytes :: ByteString -> NetworkOption

Set the ca bundle

tlsCaPath :: String -> NetworkOption

Set the file from which to load the certificate authority bundle

tlsPassword :: String -> NetworkOption

Set the passphrase for encrypted private key. Password should be set before setting the key for the password to be used.

disableMultiVersionClientApi :: NetworkOption

Disables the multi-version client API and instead uses the local client directly. Must be set before setting up the network.

callbacksOnExternalThreads :: NetworkOption

If set, callbacks from external client libraries can be called from threads created by the FoundationDB client library. Otherwise, callbacks will be called from either the thread used to add the callback or the network thread. Setting this option can improve performance when connected using an external client, but may not be safe to use in all environments. Must be set before setting up the network. WARNING: This feature is considered experimental at this time.

externalClientLibrary :: String -> NetworkOption

Adds an external client library for use by the multi-version client API. Must be set before setting up the network.

externalClientDirectory :: String -> NetworkOption

Searches the specified path for dynamic libraries and adds them to the list of client libraries for use by the multi-version client API. Must be set before setting up the network.

disableLocalClient :: NetworkOption

Prevents connections through the local client, allowing only connections through externally loaded client libraries.

clientThreadsPerVersion :: Int -> NetworkOption

Spawns multiple worker threads for each version of the client that is loaded. Setting this to a number greater than one implies disable_local_client.

disableClientStatisticsLogging :: NetworkOption

Disables logging of client statistics, such as sampled transaction activity.

enableSlowTaskProfiling :: NetworkOption

Deprecated: Deprecated in FDB C API

Deprecated

enableRunLoopProfiling :: NetworkOption

Enables debugging feature to perform run loop profiling. Requires trace logging to be enabled. WARNING: this feature is not recommended for use in production.

clientBuggifyEnable :: NetworkOption

Enable client buggify - will make requests randomly fail (intended for client testing)

clientBuggifyDisable :: NetworkOption

Disable client buggify

clientBuggifySectionActivatedProbability :: Int -> NetworkOption

Set the probability of a CLIENT_BUGGIFY section being active for the current execution.

clientBuggifySectionFiredProbability :: Int -> NetworkOption

Set the probability of an active CLIENT_BUGGIFY section being fired. A section will only fire if it was activated

distributedClientTracer :: String -> NetworkOption

Set a tracer to run on the client. Should be set to the same value as the tracer set on the server.

supportedClientVersions :: String -> NetworkOption

This option is set automatically to communicate the list of supported clients to the active client.

externalClient :: NetworkOption

This option is set automatically on all clients loaded externally using the multi-version API.

externalClientTransportId :: Int -> NetworkOption

This option tells a child on a multiversion client what transport ID to use.