FoundationDB.Layer.Subspace

Documentation

newtype Subspace

Represents a subspace of Tuple keys. A subspace is just a common prefix for a set of tuples.

Constructors

Subspace

Fields

rawPrefix :: ByteString

subspace

Arguments

:: [Elem]
Tuple with which to prefix the subspace. May not contain incomplete version stamps.

-> Subspace

Create a subspace from a tuple.

prefixedSubspace

Arguments

:: ByteString
prefix

-> [Elem]

-> Subspace

Create a subspace from a raw bytestring prefix and a tuple.

subspaceKey :: Subspace -> ByteString

extend :: Subspace -> [Elem] -> Subspace

Create a subsubspace by extending the prefix of a subspace by the given tuple.

pack :: Subspace -> [Elem] -> ByteString

unpack :: Subspace -> ByteString -> Either String [Elem]

contains

Arguments

:: Subspace

-> ByteString
encoded key

-> Bool

Returns True iff the subspace contains the given key.

subspaceRange :: Subspace -> Range

getLast :: Subspace -> Transaction (Maybe (ByteString, ByteString))

Get the last key,value pair in the subspace, if it exists.